requirements.txt
# Python
# Python==3.11.3 over 3.8 is ok

torch~=2.5.1
torchvision~=0.20.1
numpy~=2.0.2
tqdm~=4.67.0
transformers==4.33.2
lightning==2.0.9
pillow==9.5.0
matplotlib~=3.9.2
nuscenes-devkit

huggingface_hub==0.25.2
diffusers==0.11.1

numpy
pandas
scipy
ftfy
regex
accelerate
open_clip_torch #maybe needed